Add support for drawing idpPnt in graph SLAM
In graph-SLAM, the anchor part of idpPnt is not stored explicetly inside the
landmark parameters. Thus, they always need to be recovered from the frame and
sensor pose through the "anchoring factor".

Note that this "anchoring factor" needs to be properly adapted when removing
frames that anchor landmarks from a robot trajectory.