Use pap2newanchors instead of measurements2pap to get parallax angle to test
I _feel_ that using pap2newanchors makes the system more consistent, since an
error on the orientation in the current frame being tested to be a landmark
anchor can change a lot the position o the landmark. In this way the error in
the initial parallax will come from the initial anchors.